pyproject.toml
[tool.poetry]
name = "mgp-qdb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.41.1"
google-cloud-firestore = "^2.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
